realestate_agent/pyproject.toml
[project]
name = "realestate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muneeb", email = "muneebbaig71@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "mem0ai>=0.1.116",
    "openai-agents>=0.2.8",
    "playwright>=1.54.0",
    "sql>=2022.4.0",
]

[project.scripts]
realestate-agent = "realestate_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
